+++ Sources/KeyboardShortcutSettingsFileStore+SectionParsers.swift
@@ -1,0 +1,33 @@
+import CmuxSettings
+import Foundation
+
+/// Settings-file section parsers for file editor and file explorer options, extracted from `KeyboardShortcutSettingsFileStore.swift`, which sits at its file-length budget.
+extension CmuxSettingsFileStore {
+    func parseFileEditorSection(
+        _ section: [String: Any],
+        sourcePath: String,
+        snapshot: inout ResolvedSettingsSnapshot
+    ) {
+        if let value = jsonBool(section["wordWrap"]) {
+            snapshot.managedUserDefaults[FilePreviewWordWrapSettings.key] = .bool(value)
+        } else if section.keys.contains("wordWrap") {
+            logInvalid("fileEditor.wordWrap", sourcePath: sourcePath)
+        }
+    }
+
+    func parseFileExplorerSection(
+        _ section: [String: Any],
+        sourcePath: String,
+        snapshot: inout ResolvedSettingsSnapshot
+    ) {
+        if let raw = jsonString(section["doubleClickAction"]) {
+            if let action = FileExplorerDoubleClickAction(rawValue: raw) {
+                snapshot.managedUserDefaults[FileExplorerDoubleClickActionSettings.key] = .string(action.rawValue)
+            } else {
+                logInvalid("fileExplorer.doubleClickAction", sourcePath: sourcePath)
+            }
+        } else if section.keys.contains("doubleClickAction") {
+            logInvalid("fileExplorer.doubleClickAction", sourcePath: sourcePath)
+        }
+    }
+}
