+++ Sources/SessionPersistence+RightSidebarTool.swift
@@ -1,0 +1,20 @@
+import Foundation
+
+/// Right-sidebar panel snapshot, extracted from `SessionPersistence.swift`, which sits at its file-length budget.
+struct SessionRightSidebarToolPanelSnapshot: Codable, Sendable {
+    var mode: RightSidebarMode?
+
+    init(mode: RightSidebarMode?) {
+        self.mode = mode
+    }
+
+    private enum CodingKeys: String, CodingKey {
+        case mode
+    }
+
+    init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        let raw = try container.decodeIfPresent(String.self, forKey: .mode)
+        self.mode = raw.flatMap { RightSidebarMode(rawValue: $0) }
+    }
+}
